Consolidate ByteArrayHelpers
