+++ AElf.Benchmark/TransactionDataGenerator.cs
@@ -2,15 +2,13 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using AElf.Common.ByteArrayHelpers;
-using AElf.Common.Extensions;
+using AElf.Common;
 using AElf.Kernel;
 using AElf.Types.CSharp;
 using Akka.Util.Internal.Collections;
 using Google.Protobuf;
 using Org.BouncyCastle.Security;
 using ServiceStack;
-using AElf.Common;
 
 namespace AElf.Benchmark
 {
