+++ AElf.CLI/AElfCliProgram.cs
@@ -17,10 +17,8 @@ using AElf.CLI.Screen;
 using AElf.CLI.Streaming;
 using AElf.CLI.Wallet;
 using AElf.CLI.Wallet.Exceptions;
-using AElf.Common.ByteArrayHelpers;
-using AElf.Common.Extensions;
-using AElf.Cryptography.ECDSA;
 using AElf.Common;
+using AElf.Cryptography.ECDSA;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using ProtoBuf;
