+++ AElf.Common/AElf.Common.csproj
@@ -16,6 +16,7 @@
     <Exec Command="protoc --proto_path=../AElf.Kernel.Types/Protobuf/Proto/ --csharp_out=./Protobuf/Generated --csharp_opt=file_extension=.g.cs ../AElf.Kernel.Types/Protobuf/Proto/common.proto" />
   </Target>
   <ItemGroup>
+    <Folder Include="ByteArrayHelpers" />
     <Folder Include="Protobuf\Generated" />
   </ItemGroup>
 </Project>
