+++ AElf.Common/Collections/BoundedByteArrayQueue.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using AElf.Common.ByteArrayHelpers;
 
 namespace AElf.Common.Collections
 {
