+++ AElf.Kernel.Tests.TestContractZero/TestContractZero.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using AElf.Common.ByteArrayHelpers;
+using AElf.Common;
 using AElf.Contracts.Genesis;
 using AElf.Sdk.CSharp;
 using AElf.Sdk.CSharp.Types;
@@ -13,7 +13,6 @@ using SharpRepository.Repository.Configuration;
 using Api = AElf.Sdk.CSharp.Api;
 using CSharpSmartContract = AElf.Sdk.CSharp.CSharpSmartContract;
 using AElf.Kernel.KernelAccount;
-using AElf.Common;
 
 // ReSharper disable once CheckNamespace
 namespace AElf.Kernel.Tests
