+++ AElf.Miner.Rpc/Client/ClientBase.cs
@@ -4,13 +4,11 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using AElf.Common.ByteArrayHelpers;
+using AElf.Common;
 using AElf.Configuration;
 using AElf.Kernel;
 using Grpc.Core;
 using NLog;
-using AElf.Common;
-using Google.Protobuf;
 
 namespace AElf.Miner.Rpc.Client
 {
