+++ AElf.Miner/ClientManager.cs
@@ -5,7 +5,6 @@ using System.Threading;
 using System.Threading.Tasks;
 using AElf.Common;
 using AElf.Common.Attributes;
-using AElf.Common.ByteArrayHelpers;
 using AElf.Configuration.Config.GRPC;
 using AElf.Cryptography.Certificate;
 using AElf.Kernel;
