+++ AElf.Network/Connection/MessageWriter.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Net.Sockets;
 using System.Threading.Tasks;
-using AElf.Common.ByteArrayHelpers;
+using AElf.Common;
 using NLog;
 
 namespace AElf.Network.Connection
