+++ AElf.Network/Peers/PeerManager.cs
@@ -6,8 +6,7 @@ using System.Net.Sockets;
 using System.Runtime.CompilerServices;
 using System.Threading.Tasks;
 using AElf.Common.Attributes;
-using AElf.Common.ByteArrayHelpers;
-using AElf.Common.Extensions;
+using AElf.Common;
 using AElf.Configuration;
 using AElf.Configuration.Config.Network;
 using AElf.Cryptography.ECDSA;
