+++ AElf.Node/AElfChain/MainchainNodeService.cs
@@ -6,8 +6,8 @@ using System.Threading.Tasks;
 using System.Xml.Serialization;
 using AElf.ChainController;
 using AElf.ChainController.TxMemPool;
+using AElf.Common;
 using AElf.Common.Attributes;
-using AElf.Common.ByteArrayHelpers;
 using AElf.Common.Enums;
 using AElf.Configuration;
 using AElf.Configuration.Config.Consensus;
@@ -15,7 +15,6 @@ using AElf.Cryptography.ECDSA;
 using AElf.Kernel;
 using AElf.Kernel.Node;
 using AElf.Miner.Miner;
-using AElf.Node.CrossChain;
 using AElf.Node.Protocol;
 using AElf.SmartContract;
 using Google.Protobuf;
