+++ AElf.Node/Protocol/NetworkManager.cs
@@ -5,10 +5,9 @@ using System.Runtime.CompilerServices;
 using System.Threading.Tasks;
 using AElf.ChainController.EventMessages;
 using AElf.ChainController.TxMemPool;
+using AElf.Common;
 using AElf.Common.Attributes;
-using AElf.Common.ByteArrayHelpers;
 using AElf.Common.Collections;
-using AElf.Common.Extensions;
 using AElf.Kernel;
 using AElf.Network;
 using AElf.Network.Connection;
