+++ AElf.Node/Protocol/PendingBlock.cs
@@ -1,12 +1,10 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using AElf.ChainController;
-using AElf.Common.ByteArrayHelpers;
-using AElf.Common.Extensions;
+using AElf.Common;
 using AElf.Kernel;
 using AElf.Network;
 using AElf.Network.Peers;
-using AElf.Common;
 
 namespace AElf.Node.Protocol
 {
