+++ AElf.Runtime.CSharp.Tests/RunnerMetadataTest.cs
@@ -2,8 +2,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using AElf.Common.ByteArrayHelpers;
 using AElf.Configuration.Config.Contract;
+using AElf.Common;
 using AElf.Kernel;
 using AElf.Kernel.Storages;
 using AElf.SmartContract;
